FRWorkerWithAlignment/requirements.txt
# FRWorkerWithAlignment dependencies
# Python 3.10–3.13

# Core inference
insightface>=0.7.3
onnxruntime>=1.16.0          # CPU; swap for onnxruntime-gpu for CUDA
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
scipy>=1.11.0

# Storage
lancedb>=0.29.0
pyarrow>=14.0.0
pandas>=2.0.0

# Web app
Flask>=3.0.0
Werkzeug>=3.0.0

# ---- Optional: 3DDFA_V2 for full 3D frontalization (yaw > 45°) ----
# Not pip-installable. See README.md for manual installation.
# Without it the PnP-homography fallback is used automatically.
